services/backend-oauth/requirements.txt
Django==5.2.1
djangorestframework==3.16.00
djangorestframework-simplejwt==5.5.0
python-dotenv==1.1.0
uvicorn==0.34.2
psycopg2-binary==2.9.10
django-extensions==4.1 
oauthlib==3.2.2
django-oauth-toolkit==1.6.3

